+++ packages/plugin-compiler/src/main/kotlin/io/realm/kotlin/compiler/Registrar.kt
@@ -98,6 +98,4 @@ class Registrar : ComponentRegistrar {
             }
         }
     }
-
-    override val supportsK2: Boolean = true
 }