+++ packages/test-base/src/jvmMain/kotlin/io/realm/kotlin/test/util/Compiler.kt
@@ -17,7 +17,6 @@
 
 package io.realm.kotlin.test.util
 
-import com.tschuchort.compiletesting.JvmCompilationResult
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import io.realm.kotlin.compiler.Registrar
@@ -28,7 +27,7 @@ object Compiler {
     fun compileFromSource(
         source: SourceFile,
         plugins: List<Registrar> = listOf(Registrar())
-    ): JvmCompilationResult =
+    ): KotlinCompilation.Result =
         KotlinCompilation().apply {
             sources = listOf(source)
             messageOutputStream = System.out