+++ packages/test-base/src/jvmTest/kotlin/io/realm/kotlin/test/compiler/Utils.kt
@@ -2,14 +2,13 @@
 
 package io.realm.kotlin.test.compiler
 
-import com.tschuchort.compiletesting.JvmCompilationResult
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import io.realm.kotlin.compiler.CollectionType
 import io.realm.kotlin.test.util.Compiler
 import org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi
 
-fun createFileAndCompile(fileName: String, code: String): JvmCompilationResult =
+fun createFileAndCompile(fileName: String, code: String): KotlinCompilation.Result =
     Compiler.compileFromSource(SourceFile.kotlin(fileName, code))
 
 /**